fixed the bugs in the blog, made a blog listing page, styled and add filter, editted the nav bar mobile view to show name and fixed the color bug in the hamburger menu. worked on the homepage and started on the footer
moved the blog folder outside pages and added an mdx folder to the lib

styled and added the navigation bar

made some changes to the blog card

added the skeleton of the customer review component

started on the footer- made the privacy policy component and page and linked it to the footer. started working on the terms of service component and page.